+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## [Unreleased]
 
 * (ibc-go) Use ibc v4.4.0 https://pkg.akt.dev/node/pull/1825
+* Pause deployment groups when account is overdrawn
 
 ### Improvements